+++ awscli/customizations/s3/utils.py
@@ -145,6 +145,34 @@ def get_file_stat(path):
     return stats.st_size, update_time
 
 
+def find_dest_path_comp_key(files, src_path=None):
+    """
+    This is a helper function that determines the destination path and compare
+    key given parameters received from the ``FileFormat`` class.
+    """
+    src = files['src']
+    dest = files['dest']
+    src_type = src['type']
+    dest_type = dest['type']
+    if src_path is None:
+        src_path = src['path']
+
+    sep_table = {'s3': '/', 'local': os.sep}
+
+    if files['dir_op']:
+        rel_path = src_path[len(src['path']):]
+    else:
+        rel_path = src_path.split(sep_table[src_type])[-1]
+    compare_key = rel_path.replace(sep_table[src_type], '/')
+    if files['use_src_name']:
+        dest_path = dest['path']
+        dest_path += rel_path.replace(sep_table[src_type],
+                                      sep_table[dest_type])
+    else:
+        dest_path = dest['path']
+    return dest_path, compare_key
+
+
 def check_etag(etag, fileobj):
     """
     This fucntion checks the etag and the md5 checksum to ensure no
@@ -244,6 +272,21 @@ def uni_print(statement, out_file=None):
             # this works with a normal StringIO.
             out_file.write(statement.encode('utf-8'))
     out_file.flush()
+
+
+def bytes_print(statement):
+    """
+    This function is used to properly write bytes to standard out.
+    """
+    if PY3:
+        if getattr(sys.stdout, 'buffer', None):
+            sys.stdout.buffer.write(statement)
+        else:
+            # If it is not possible to write to the standard out buffer.
+            # The next best option is to decode and write to standard out.
+            sys.stdout.write(statement.decode('utf-8'))
+    else:
+        sys.stdout.write(statement)
 
 
 def guess_content_type(filename):
@@ -404,7 +447,8 @@ class PrintTask(namedtuple('PrintTask',
                                              warning)
 
 
-IORequest = namedtuple('IORequest', ['filename', 'offset', 'data'])
+IORequest = namedtuple('IORequest',
+                       ['filename', 'offset', 'data', 'is_stream'])
 # Used to signal that IO for the filename is finished, and that
 # any associated resources may be cleaned up.
 IOCloseRequest = namedtuple('IOCloseRequest', ['filename'])