+++ tests/unit/customizations/s3/test_subcommands.py
@@ -174,12 +174,13 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                         'rb': ['s3_handler']}
 
         params = {'filters': True, 'region': 'us-east-1', 'endpoint_url': None,
-                  'verify_ssl': None}
+                  'verify_ssl': None, 'is_stream': False}
         for cmd in cmds:
             cmd_arc = CommandArchitecture(self.session, cmd,
                                           {'region': 'us-east-1',
                                            'endpoint_url': None,
-                                           'verify_ssl': None})
+                                           'verify_ssl': None,
+                                           'is_stream': False})
             cmd_arc.create_instructions()
             self.assertEqual(cmd_arc.instructions, instructions[cmd])
 
@@ -202,7 +203,8 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': local_file, 'dest': s3_file, 'filters': filters,
                   'paths_type': 'locals3', 'region': 'us-east-1',
                   'endpoint_url': None, 'verify_ssl': None,
-                  'follow_symlinks': True, 'page_size': None}
+                  'follow_symlinks': True, 'page_size': None,
+                  'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'cp', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -218,7 +220,8 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': local_file, 'dest': s3_file, 'filters': filters,
                   'paths_type': 'locals3', 'region': 'us-east-1',
                   'endpoint_url': None, 'verify_ssl': None,
-                  'follow_symlinks': True, 'page_size': None}
+                  'follow_symlinks': True, 'page_size': None,
+                  'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'cp', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -241,7 +244,8 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': s3_file, 'dest': local_file, 'filters': filters,
                   'paths_type': 's3local', 'region': 'us-east-1',
                   'endpoint_url': None, 'verify_ssl': None,
-                  'follow_symlinks': True, 'page_size': None}
+                  'follow_symlinks': True, 'page_size': None,
+                  'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'cp', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -258,7 +262,8 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': s3_file, 'dest': s3_file, 'filters': filters,
                   'paths_type': 's3s3', 'region': 'us-east-1',
                   'endpoint_url': None, 'verify_ssl': None,
-                  'follow_symlinks': True, 'page_size': None}
+                  'follow_symlinks': True, 'page_size': None,
+                  'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'cp', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -275,7 +280,8 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': s3_file, 'dest': s3_file, 'filters': filters,
                   'paths_type': 's3s3', 'region': 'us-east-1',
                   'endpoint_url': None, 'verify_ssl': None,
-                  'follow_symlinks': True, 'page_size': None}
+                  'follow_symlinks': True, 'page_size': None,
+                  'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'mv', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -292,7 +298,8 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': s3_file, 'dest': s3_file, 'filters': filters,
                   'paths_type': 's3', 'region': 'us-east-1',
                   'endpoint_url': None, 'verify_ssl': None,
-                  'follow_symlinks': True, 'page_size': None}
+                  'follow_symlinks': True, 'page_size': None,
+                  'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'rm', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -313,7 +320,8 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': local_dir, 'dest': s3_prefix, 'filters': filters,
                   'paths_type': 'locals3', 'region': 'us-east-1',
                   'endpoint_url': None, 'verify_ssl': None,
-                  'follow_symlinks': True, 'page_size': None}
+                  'follow_symlinks': True, 'page_size': None,
+                  'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'sync', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -329,7 +337,7 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': s3_prefix, 'dest': s3_prefix, 'paths_type': 's3',
                   'region': 'us-east-1', 'endpoint_url': None,
                   'verify_ssl': None, 'follow_symlinks': True,
-                  'page_size': None}
+                  'page_size': None, 'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'mb', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -345,7 +353,7 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': s3_prefix, 'dest': s3_prefix, 'paths_type': 's3',
                   'region': 'us-east-1', 'endpoint_url': None,
                   'verify_ssl': None, 'follow_symlinks': True,
-                  'page_size': None}
+                  'page_size': None, 'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'rb', params)
         cmd_arc.create_instructions()
         rc = cmd_arc.run()
@@ -362,7 +370,7 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': s3_prefix, 'dest': s3_prefix, 'paths_type': 's3',
                   'region': 'us-east-1', 'endpoint_url': None,
                   'verify_ssl': None, 'follow_symlinks': True,
-                  'page_size': None}
+                  'page_size': None, 'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'rb', params)
         cmd_arc.create_instructions()
         rc = cmd_arc.run()
@@ -472,6 +480,34 @@ class CommandParametersTest(unittest.TestCase):
         cmd_params = CommandParameters(self.session, 'rb', {'force': True},'')
         cmd_params.parameters['src'] = 's3://mybucket'
         cmd_params.check_force(None)
+
+    def test_validate_streaming_paths_upload(self):
+        parameters = {'src': '-', 'dest': 's3://bucket'}
+        cmd_params = CommandParameters(self.session, 'cp', parameters, '')
+        cmd_params._validate_streaming_paths()
+        self.assertTrue(cmd_params.parameters['is_stream'])
+        self.assertTrue(cmd_params.parameters['only_show_errors'])
+        self.assertFalse(cmd_params.parameters['dir_op'])
+
+    def test_validate_streaming_paths_download(self):
+        parameters = {'src': 'localfile', 'dest': '-'}
+        cmd_params = CommandParameters(self.session, 'cp', parameters, '')
+        cmd_params._validate_streaming_paths()
+        self.assertTrue(cmd_params.parameters['is_stream'])
+        self.assertTrue(cmd_params.parameters['only_show_errors'])
+        self.assertFalse(cmd_params.parameters['dir_op'])
+
+    def test_validate_no_streaming_paths(self):
+        parameters = {'src': 'localfile', 'dest': 's3://bucket'}
+        cmd_params = CommandParameters(self.session, 'cp', parameters, '')
+        cmd_params._validate_streaming_paths()
+        self.assertFalse(cmd_params.parameters['is_stream'])
+
+    def test_validate_streaming_paths_error(self):
+        parameters = {'src': '-', 'dest': 's3://bucket'}
+        cmd_params = CommandParameters(self.session, 'sync', parameters, '')
+        with self.assertRaises(ValueError):
+            cmd_params._validate_streaming_paths()
 
 
 class HelpDocTest(BaseAWSHelpOutputTest):